sales entry

+++ Medicine/RemoveMedicine.cs
@@ -49,19 +49,44 @@ namespace MedicalStoreManagementSystem.Medicine
         }
 
         private void button_removeMedicine_Click(object sender, EventArgs e)
-        { 
+        {
+            db.closeConnection();
             if (comboBox_medicine.Text != "")
             {
                 try
                 {
-                    query = $"DELETE FROM medicine WHERE medicineName = '{comboBox_medicine.Text}'";
-                    db.setData(query);
-                    MessageBox.Show("Medicine Deleted");
 
-                    comboBox_medicine.Items.Clear();
-                    Load_Medicine();
-                    comboBox_medicine.Text = string.Empty;
-                    comboBox_medicine.Focus();
+                    query = $"SELECT quantity FROM stockRecord WHERE medicineName = '{comboBox_medicine.Text}'";
+                    dr = db.getData(query);
+                    dr.Read();
+                    int quantity = Convert.ToInt32(dr[0].ToString());
+                    db.closeConnection();
+
+                    if (quantity > 0)
+                    {
+                        MessageBox.Show("Medicine Available In Stock, first Remove from Stock");
+                    }
+                    else
+                    {
+
+
+                        if (quantity == 0)
+                        {
+                            query = $"DELETE FROM stockRecord WHERE medicineName = '{comboBox_medicine.Text}'";
+                            db.setData(query);
+
+                            query = $"DELETE FROM medicine WHERE medicineName = '{comboBox_medicine.Text}'";
+                            db.setData(query);
+
+                            MessageBox.Show("Medicine Deleted");
+                        }
+
+
+                        comboBox_medicine.Items.Clear();
+                        Load_Medicine();
+                        comboBox_medicine.Text = string.Empty;
+                        comboBox_medicine.Focus();
+                    }
                 }
                 catch (Exception ex)
                 {
